Move labeling classes out of QgsMapRenderer header
