+++ python/core/qgsmaprenderer.sip
@@ -1,78 +1,4 @@
 
-
-class QgsLabelPosition
-{
-%TypeHeaderCode
-#include <qgsmaprenderer.h>
-%End
-  public:
-    QgsLabelPosition( int id, double r, const QVector< QgsPoint >& corners, const QgsRectangle& rect, double w, double h, const QString& layer, const QString& labeltext, const QFont& labelfont, bool upside_down, bool diagram = false, bool pinned = false, const QString& providerId = QString() );
-    QgsLabelPosition();
-    int featureId;
-    double rotation;
-    QVector< QgsPoint > cornerPoints;
-    QgsRectangle labelRect;
-    double width;
-    double height;
-    QString layerID;
-    QString labelText;
-    QFont labelFont;
-    bool upsideDown;
-    bool isDiagram;
-    bool isPinned;
-    //! @note added in 2.14
-    QString providerID;
-};
-
-/** Labeling engine interface. */
-class QgsLabelingEngineInterface
-{
-%TypeHeaderCode
-#include <qgsmaprenderer.h>
-%End
-
-  public:
-
-    virtual ~QgsLabelingEngineInterface();
-
-    //! called when we're going to start with rendering
-    virtual void init( const QgsMapSettings& mapSettings ) = 0;
-    //! called to find out whether the layer is used for labeling
-    virtual bool willUseLayer( QgsVectorLayer* layer ) = 0;
-    //! clears all PAL layer settings for registered layers
-    virtual void clearActiveLayers() = 0;
-    //! clears data defined objects from PAL layer settings for a registered layer
-    virtual void clearActiveLayer( const QString& layerID ) = 0;
-    //! called when starting rendering of a layer
-    virtual int prepareLayer( QgsVectorLayer* layer, QStringList& attrNames, QgsRenderContext& ctx ) = 0;
-    //! returns PAL layer settings for a registered layer
-    //! @deprecated since 2.12 - if direct access to QgsPalLayerSettings is necessary, use QgsPalLayerSettings::fromLayer()
-    virtual QgsPalLayerSettings &layer( const QString &layerName ) = 0 /Deprecated/;
-    //! adds a diagram layer to the labeling engine
-    //! @note added in QGIS 2.12
-    virtual int prepareDiagramLayer( QgsVectorLayer *layer, QStringList &attrNames, QgsRenderContext &ctx );
-    //! adds a diagram layer to the labeling engine
-    //! @deprecated since 2.12 - use prepareDiagramLayer()
-    virtual int addDiagramLayer( QgsVectorLayer *layer, const QgsDiagramLayerSettings *s ) /Deprecated/;
-    //! called for every feature
-    virtual void registerFeature( const QString &layerID, QgsFeature &feat, QgsRenderContext &context ) = 0;
-    //! called for every diagram feature
-    virtual void registerDiagramFeature( const QString &layerID, QgsFeature &feat, QgsRenderContext &context );
-    //! called when the map is drawn and labels should be placed
-    virtual void drawLabeling( QgsRenderContext& context ) = 0;
-    //! called when we're done with rendering
-    virtual void exit() = 0;
-    //! return infos about labels at a given (map) position
-    //! @deprecated since 2.4 - use takeResults() and methods of QgsLabelingResults
-    virtual QList<QgsLabelPosition> labelsAtPosition( const QgsPoint& p ) = 0 /Deprecated/;
-    //! return infos about labels within a given (map) rectangle
-    //! @deprecated since 2.4 - use takeResults() and methods of QgsLabelingResults
-    virtual QList<QgsLabelPosition> labelsWithinRect( const QgsRectangle& r ) = 0 /Deprecated/;
-
-    //! called when passing engine among map renderers
-    virtual QgsLabelingEngineInterface* clone() = 0 /Factory/;
-};
-
 struct QgsLayerCoordinateTransform
 {
   QString srcAuthId;
